Move check_isallfinite to utils
